Rename pds.web_analytics Python package to pds.o11y_batch
Repo/Terraform naming already used o11y-cloudfront-batch, but the actual Python package (src/pds/web_analytics), its imports/mock targets in tests, docs, and misc references were still fully on the old web_analytics/web-analytics naming. Renamed the package to pds.o11y_batch and updated all consumers (setup.cfg, MANIFEST.in, scripts, tests, README/CLAUDE, docs/source/*.rst, docker comments, .detect-secrets-ignore). CLI entry point commands (s3-log-sync, egress-report) are unchanged, only their underlying module path. CHANGELOG.md, docs/requirements/** (generated), and archive/sql/** (references real historical bucket paths) intentionally left untouched. Also fixed a stale terraform/README.md instruction referencing a pds-main tfvar that no longer exists (actual var is enable_o11y_batch). Full test suite passes against the renamed package (47 passed, 5 skipped).
