+++ .detect-secrets-ignore
@@ -1,4 +1,4 @@
-# Per-repo secret detection exclusions for pds-web-analytics.
+# Per-repo secret detection exclusions for pds-o11y-cloudfront-batch.
 # One regex per line; lines starting with # are comments.
 # These supplement the global exclusions in scripts/detect_secrets_baseline.sh.
