+++ scripts/egress_report.py
@@ -6,7 +6,7 @@ This script is maintained for backward compatibility.
 For new installations, use the 'egress-report' command directly.
 """
 
-from pds.web_analytics.egress_report import main
+from pds.o11y_batch.egress_report import main
 
 if __name__ == "__main__":
     main()
