+++ setup.cfg
@@ -23,7 +23,7 @@ author_email = pds_operator@jpl.nasa.gov
 description = PDS Web Analytics Batch - Log processing and S3 synchronization for Planetary Data System
 long_description = file: README.md
 long_description_content_type = text/markdown
-version = file: src/pds/web_analytics/VERSION.txt
+version = file: src/pds/o11y_batch/VERSION.txt
 license = apache-2.0
 keywords = pds, planetary data, web analytics, logstash, s3, aws
 url = https://github.com/NASA-PDS/o11y-cloudfront-batch
@@ -88,8 +88,8 @@ dev =
 [options.entry_points]
 # Put your entry point scripts here
 console_scripts =
-    s3-log-sync = pds.web_analytics.s3_sync:main
-    egress-report = pds.web_analytics.egress_report:main
+    s3-log-sync = pds.o11y_batch.s3_sync:main
+    egress-report = pds.o11y_batch.egress_report:main
 
 [options.packages.find]
 # Don't change this. Needed to find packages under src/
